Quick changes from merge to reorder components and rename blog component
