+++ landing-page/src/App.tsx
@@ -5,7 +5,7 @@ import AboutUs from './components/layout/AboutUs';
 import styled from 'styled-components';
 import FooterBar from './components/layout/FooterBar';
 import Community from './components/layout/Community';
-import LitePaper from './components/layout/LitePaper';
+import Blog from './components/layout/Blog';
 
 function App() {
   return (
@@ -14,8 +14,8 @@ function App() {
       <MainSection />
       <PartnerLogos />
       <AboutUs />
+      <Blog/>
       <Community />
-      <LitePaper />
       <Div />
       <FooterBar />
     </LandingPage>
