+++ landing-page/src/components/layout/Blog.tsx
