+++ landing-page/src/components/layout/FooterBar.tsx
@@ -4,7 +4,7 @@ import logo from '../../assets/horizontal_logo_orange&white.svg'
 
 function FooterBar() {
   return (
-    <Footer id='community' >
+    <Footer >
       <Img src={logo} alt={"Y-Foundry Logo"} />
       <SocialIcons />
     </Footer>
